fix 6-size,c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,11 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-	printf("Size of a char: %d byte(s)", (int)sizeof(a));
-	Printf("Size of an int: %d byte(s)", (int)sizeof(b));
-	printf("Size of a long int: %d byte(s)", (int)sizeof(c));
-	printf("Size of a long long int: %d byte(s)", (int)sizeof(d));
-	printf("Size of a float: %d byte(s)", (int)sizeof(e));
+	printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(a));
+	Printf("Size of an int: %lu byte(s)", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu byte(s)", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(e));
+	
 	return (0);
 }
